+++ test/entrypoints.test.tsx
@@ -1,3 +1,4 @@
+import { fromAbsolute } from "@internationalized/date";
 import { Temporal as JsTemporal } from "@js-temporal/polyfill";
 import dayjs from "dayjs";
 import { DateTime } from "luxon";
@@ -12,6 +13,7 @@ import * as native from "../src";
 import type { ConfiguredNeodtProps, ConfiguredNaturalDateParseOptions } from "../src/configured";
 import * as dates from "../src/date-fns";
 import * as days from "../src/dayjs";
+import * as internationalized from "../src/internationalized-date";
 import * as jsTemporal from "../src/js-temporal-polyfill";
 import * as luxon from "../src/luxon";
 import * as moments from "../src/moment";
@@ -69,6 +71,12 @@ function contract<T, TZone>(
   });
 }
 
+contract(
+  "@internationalized/date",
+  internationalized,
+  fromAbsolute(milliseconds, "America/New_York"),
+  (value) => value.toDate().getTime(),
+);
 contract("luxon", luxon, DateTime.fromMillis(milliseconds), (value) => value.toMillis());
 contract("moment", moments, moment(milliseconds), (value) => value.valueOf());
 contract("dayjs", days, dayjs(milliseconds), (value) => value.valueOf());